usuario/container/requirements.txt
anyio==3.6.1
bcrypt==4.0.0
borneo==5.3.5
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
circuitbreaker==1.4.0
click==8.1.3
cryptography==37.0.2
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.85.0
h11==0.13.0
idna==3.4
oci==2.82.0
passlib==1.7.4
pycparser==2.21
pydantic==1.10.2
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
typing-extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
python-jose==3.3.0
